+++ src/oauth2Auth.js
@@ -38,6 +38,14 @@ class Oauth2Auth extends AlfrescoApiClient {
                 this.config.oauth2.refreshTokenTimeout = 40000;
             }
 
+            if (!this.config.oauth2.redirectSilentIframeUri) {
+                var context = '';
+                if (typeof window !== 'undefined') {
+                    context = window.location.origin;
+                }
+                this.config.oauth2.redirectSilentIframeUri = context + '/assets/silent-refresh.html';
+            }
+
             this.basePath = this.config.oauth2.host; //Auth Call
 
             this.authentications = {
@@ -57,6 +65,10 @@ class Oauth2Auth extends AlfrescoApiClient {
     }
 
     initOauth() {
+        if (!this.config.oauth2.implicitFlow && this.isValidAccessToken()) {
+            const accessToken = this.storage.getItem('access_token');
+            this.setToken(accessToken, null);
+        }
         return Promise.resolve()
             .then(() => {
                 return this.discoveryUrls();
@@ -157,9 +169,10 @@ class Oauth2Auth extends AlfrescoApiClient {
 
             this.hashFragmentParams = this.getHashFragmentParams(externalHash);
 
-            if (this.isValidAccessToken()) {
+            if (externalHash === undefined && this.isValidAccessToken()) {
                 let accessToken = this.storage.getItem('access_token');
                 this.setToken(accessToken, null);
+                this.silentRefresh();
                 resolve(accessToken);
             }
 
@@ -169,7 +182,6 @@ class Oauth2Auth extends AlfrescoApiClient {
                 let sessionState = this.hashFragmentParams.session_state;
                 let expiresIn = this.hashFragmentParams.expires_in;
 
-                window.location.hash = '';
                 if (!sessionState) {
                     reject('session state not present');
                 }
@@ -364,12 +376,35 @@ class Oauth2Auth extends AlfrescoApiClient {
 
     }
 
+    composeIframeLoginUrl() {
+        var nonce = this.genNonce();
+
+        this.storage.setItem('nonce', nonce);
+
+        var separation = this.discovery.loginUrl.indexOf('?') > -1 ? '&' : '?';
+
+        return this.discovery.loginUrl +
+            separation +
+            'client_id=' +
+            encodeURIComponent(this.config.oauth2.clientId) +
+            '&redirect_uri=' +
+            encodeURIComponent(this.config.oauth2.redirectSilentIframeUri) +
+            '&scope=' +
+            encodeURIComponent(this.config.oauth2.scope) +
+            '&response_type=' +
+            encodeURIComponent('id_token token') +
+            '&nonce=' +
+            encodeURIComponent(nonce) +
+            '&prompt=none';
+
+    }
+
     hasHashCharacter(hash) {
         return hash.indexOf('#') === 0;
     }
 
-    isHashRoute(hash) {
-        return hash.indexOf('#/') === 0;
+    startWithHashRoute(hash) {
+        return hash.startsWith('#/');
     }
 
     getHashFragmentParams(externalHash) {
@@ -380,11 +415,16 @@ class Oauth2Auth extends AlfrescoApiClient {
 
             if (!externalHash) {
                 hash = decodeURIComponent(window.location.hash);
+                if (!this.startWithHashRoute(hash)) {
+                    window.location.hash = '';
+                }
             } else {
                 hash = decodeURIComponent(externalHash);
+                this.removeHashFromSilentIframe();
+                this.destroyIframe();
             }
 
-            if (this.hasHashCharacter(hash) && !this.isHashRoute(hash)) {
+            if (this.hasHashCharacter(hash) && !this.startWithHashRoute(hash)) {
                 const questionMarkPosition = hash.indexOf('?');
 
                 if (questionMarkPosition > -1) {
@@ -442,10 +482,17 @@ class Oauth2Auth extends AlfrescoApiClient {
         }, this.config.oauth2.refreshTokenTimeout);
     }
 
+    removeHashFromSilentIframe() {
+        var iframe = document.getElementById('silent_refresh_token_iframe');
+        if (iframe && iframe.contentWindow.location.hash) {
+            iframe.contentWindow.location.hash = '';
+        }
+    }
+
     createIframe() {
         const iframe = document.createElement('iframe');
         iframe.id = 'silent_refresh_token_iframe';
-        let loginUrl = this.composeImplicitLoginUrl();
+        let loginUrl = this.composeIframeLoginUrl();
         iframe.setAttribute('src', loginUrl);
         iframe.style.display = 'none';
         document.body.appendChild(iframe);
@@ -517,6 +564,7 @@ class Oauth2Auth extends AlfrescoApiClient {
             (data) => {
                 this.saveUsername(username);
                 this.setToken(data.access_token, data.refresh_token);
+                this.storeAccessToken(data.access_token, data.expires_in);
                 resolve(data);
             },
             (error) => {
@@ -647,9 +695,13 @@ class Oauth2Auth extends AlfrescoApiClient {
             '&id_token_hint=' +
             encodeURIComponent(id_token);
 
-        if (this.config.oauth2.implicitFlow && typeof window !== 'undefined') {
-            window.location.href = logoutUrl;
-        }
+        var returnPromise =  Promise.resolve().then(() => {
+            if (this.config.oauth2.implicitFlow && typeof window !== 'undefined') {
+                window.location.href = logoutUrl;
+            }
+        });
+
+        return returnPromise;
     }
 
     invalidateSession() {